chore: moves AnonymousComponent to XAnonymous
