+++ packages/kuma-gui/src/app/x/components/x-anonymous/XAnonymous.vue
