+++ packages/kuma-gui/src/app/x/index.ts
@@ -4,6 +4,7 @@ import XAboutCard from './components/x-about-card/XAboutCard.vue'
 import XAction from './components/x-action/XAction.vue'
 import XActionGroup from './components/x-action-group/XActionGroup.vue'
 import XAlert from './components/x-alert/XAlert.vue'
+import XAnonymous from './components/x-anonymous/XAnonymous.vue'
 import XBadge from './components/x-badge/XBadge.vue'
 import XBreadcrumbs from './components/x-breadcrumbs/XBreadcrumbs.vue'
 import XCodeBlock from './components/x-code-block/XCodeBlock.vue'
@@ -38,6 +39,7 @@ declare module 'vue' {
     XTooltip: typeof KTooltip
     //
     XAlert: typeof XAlert
+    XAnonymous: typeof XAnonymous
     XIcon: typeof XIcon
     XI18n: typeof XI18n
     XInput: typeof XInput
@@ -91,6 +93,7 @@ export const services = (app: Record<string, Token>): ServiceDefinition[] => {
           //
           ['XAction', XAction],
           ['XActionGroup', XActionGroup],
+          ['XAnonymous', XAnonymous],
           ['XBadge', XBadge],
           ['XBreadcrumbs', XBreadcrumbs],
           ['XCopyButton', XCopyButton],
